+++ crates/tycho-execution/src/encoding/evm/swap_encoder/mod.rs
@@ -18,6 +18,7 @@ mod lunarbase;
 mod maverick_v2;
 mod metric;
 mod native_wrap;
+mod propamm;
 mod ring_swap_v2;
 mod rocketpool;
 mod sky;